create_citations

This function creates citations for multiple R packages at once. By default, this function will create a citation for R, along with the other functions you include. When creating the citations, you can choose to either receive text citations or have them formatted in BibTeX. The function will return citations in the console pane in your RStudio window.

create_citations(packages = NULL, format = "bibtex", include_r = TRUE)

Arguments

packages
A list of packages that you have used or plan to use in your document. If packages argument is empty, the function will create citations for packages loaded into current R session.

format
The type of citations you want to return. Returns bibtex by default, use `format = text` to return text citations.

include_r
If set to TRUE, will create a citation for R.

Value

A list of citations for the packages included in a document.

Examples

create_citations(packages = list("tidyverse", "DoSStoolkit"),
 format = "bibtex")

#> @Article{,
#>   title = {Welcome to the {tidyverse}},
#>   author = {Hadley Wickham and Mara Averick and Jennifer Bryan and Winston Chang and Lucy D'Agostino McGowan and Romain François and Garrett Grolemund and Alex Hayes and Lionel Henry and Jim Hester and Max Kuhn and Thomas Lin Pedersen and Evan Miller and Stephan Milton Bache and Kirill Müller and Jeroen Ooms and David Robinson and Dana Paige Seidel and Vitalie Spinu and Kohske Takahashi and Davis Vaughan and Claus Wilke and Kara Woo and Hiroaki Yutani},
#>   year = {2019},
#>   journal = {Journal of Open Source Software},
#>   volume = {4},
#>   number = {43},
#>   pages = {1686},
#>   doi = {10.21105/joss.01686},
#> }
#> 
#> @Manual{,
#>   title = {DoSStoolkit: Self-Paced Modules to Help You Learn R},
#>   author = {Rohan Alexander and Samantha-Jo Caetano and Annie Collins and Haoluan Chen and Isaac Ehrlich and Mariam Walaa and Marija Pejcinovska and Matthew Wankiewicz and Michael Chong and Paul Hodgetts and Shirley Deng and Yena Joo},
#>   year = {2021},
#>   note = {R package version 0.2.0},
#>   url = {https://github.com/RohanAlexander/DoSStoolkit},
#> }
#> 
#> @Manual{,
#>   title = {R: A Language and Environment for Statistical Computing},
#>   author = {{R Core Team}},
#>   organization = {R Foundation for Statistical Computing},
#>   address = {Vienna, Austria},
#>   year = {2021},
#>   url = {https://www.R-project.org/},
#> }

create_citations(list("Lahman", "ggplot2", "forcats"),
"text")

#> Friendly M, Dalzell C, Monkman M, Murphy D (2021). _Lahman: Sean
#> 'Lahman' Baseball Database_. R package version 9.0-0, <URL:
#> https://CRAN.R-project.org/package=Lahman>.
#> 
#> Wickham H (2016). _ggplot2: Elegant Graphics for Data Analysis_.
#> Springer-Verlag New York. ISBN 978-3-319-24277-4, <URL:
#> https://ggplot2.tidyverse.org>.
#> 
#> Wickham H (2021). _forcats: Tools for Working with Categorical
#> Variables (Factors)_. R package version 0.5.1, <URL:
#> https://CRAN.R-project.org/package=forcats>.
#> 
#> R Core Team (2021). _R: A Language and Environment for Statistical
#> Computing_. R Foundation for Statistical Computing, Vienna, Austria.
#> <URL: https://www.R-project.org/>.

create_citations(list("tidyverse"))

#> @Article{,
#>   title = {Welcome to the {tidyverse}},
#>   author = {Hadley Wickham and Mara Averick and Jennifer Bryan and Winston Chang and Lucy D'Agostino McGowan and Romain François and Garrett Grolemund and Alex Hayes and Lionel Henry and Jim Hester and Max Kuhn and Thomas Lin Pedersen and Evan Miller and Stephan Milton Bache and Kirill Müller and Jeroen Ooms and David Robinson and Dana Paige Seidel and Vitalie Spinu and Kohske Takahashi and Davis Vaughan and Claus Wilke and Kara Woo and Hiroaki Yutani},
#>   year = {2019},
#>   journal = {Journal of Open Source Software},
#>   volume = {4},
#>   number = {43},
#>   pages = {1686},
#>   doi = {10.21105/joss.01686},
#> }
#> 
#> @Manual{,
#>   title = {R: A Language and Environment for Statistical Computing},
#>   author = {{R Core Team}},
#>   organization = {R Foundation for Statistical Computing},
#>   address = {Vienna, Austria},
#>   year = {2021},
#>   url = {https://www.R-project.org/},
#> }